requirements.txt
python-telegram-bot
pandas
matplotlib
python-dotenv